% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{addAxes}
\alias{addAxes}
\title{addAxes}
\usage{
addAxes(lwd = 1, showXlab = T, showYlab = T, showXAxis = T,
  showYAxis = T, cex.axis = 1)
}
\arguments{
\item{lwd}{}

\item{showXlab}{}

\item{showYlab}{}

\item{showXAxis}{}

\item{showYAxis}{}

\item{cex.axis}{}
}
\description{
This function add the access to the plot
}

