% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{addYLab}
\alias{addYLab}
\title{addYLab
add an y-axis label shifted by marginShift from par()$mgp[1]}
\usage{
addYLab(ylab = "", marginShift = 0.3,
  cex.lab = LFQbench.Config$AxisLabelSize, ...)
}
\arguments{
\item{...}{other parameter passed to title(...)}

\item{ylab=""}{}

\item{marginShift=-0.2}{}

\item{cex.lab=LFQbench.Config$AxisLabelSize}{}
}
\description{
addYLab
add an y-axis label shifted by marginShift from par()$mgp[1]
}

