% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{avg_top_n}
\alias{avg_top_n}
\title{sum_top_n}
\usage{
avg_top_n(values, n, minimum = 1)
}
\arguments{
\item{values}{the vector with the values.}

\item{n}{number of top N values to be averaged.}

\item{minimum}{minimum number of valid (not NA) values allowed in order to return a value.}
}
\description{
This function averages the values of the top N values of a vector 'values'. It also allows
to have a minimum number of values to be averaged. If the minimum of values is not met, it
returns NA.
}
\details{
This top N approach is INDIVIDUAL, that is, there is no consensus among replicates to
choose the top N peptides.
}

