% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{avgNA}
\alias{avgNA}
\title{avgNA}
\usage{
avgNA(values)
}
\arguments{
\item{values}{}
}
\description{
averages values substituting NAs by zeroes
This is useful for summary functions in dplyr, for example
}

