% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.R
\name{benchLFQExperiment}
\alias{benchLFQExperiment}
\title{benchLFQExperiment}
\usage{
benchLFQExperiment(projectDir, outputDir)
}
\arguments{
\item{projectDir}{A string specifying the input directory}

\item{ouptDir}{A string specifying the output directory}
}
\description{
Analyze complete experiment from original folder and copy the results to the output folder
}

