% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.cfg.r
\name{changeConfig}
\alias{changeConfig}
\title{changeConfig}
\usage{
changeConfig(DataRootFolder = cfg$DataRootFolder,
  SampleComposition = cfg$SampleComposition,
  BackgroundSpeciesName = cfg$BackgroundSpeciesName,
  InputExtensionPattern = cfg$InputExtensionPattern,
  CsvColumnSeparator = cfg$CsvColumnSeparator,
  CsvDecimalPointChar = cfg$CsvDecimalPointChar,
  BoxPlotWhiskerQuantile = cfg$BoxPlotWhiskerQuantile,
  MinProteinAmount = cfg$MinProteinAmount,
  DropInvalidLogRatio = cfg$DropInvalidLogRatio,
  LogRatioValidityRange = cfg$LogRatioValidityRange,
  LogRatioPlotRange = cfg$LogRatioPlotRange,
  LogIntensityPlotRange = cfg$LogIntensityPlotRange,
  MaxLogRatioForAUQC = cfg$MaxLogRatioForAUQC,
  Log2IntensityRanges = cfg$Log2IntensityRanges,
  CenterLogRatioByBackground = cfg$CenterLogRatioByBackground,
  NormalizeAmountsToPPM = cfg$NormalizeAmountsToPPM,
  PlotWidth = cfg$PlotWidth, PlotHeight = cfg$PlotHeight,
  PlotCurveLineWidth = cfg$PlotCurveLineWidth,
  PlotLegendLineWidth = cfg$PlotLegendLineWidth,
  PlotPointSize = cfg$PlotPointSize,
  ScatterPlotPointType = cfg$ScatterPlotPointType,
  PlotPointMinAlpha = cfg$PlotPointMinAlpha,
  AxisLabelSize = cfg$AxisLabelSize,
  AxisAnnotationSize = cfg$AxisAnnotationSize,
  AxisLineThickness = cfg$AxisLineThickness, par = cfg$par)
}
\arguments{
\item{DataRootFolder}{the parent folder for batch processing}

\item{SampleComposition}{the quantitative composition of hybrid proteome samples}

\item{BackgroundSpeciesName}{the name of non regulated background species}

\item{InputExtensionPattern}{the pattern of file extension for batch input files}

\item{CsvColumnSeparator}{the input file column delimiter character}

\item{CsvDecimalPointChar}{the input file decimal point character}

\item{BoxPlotWhiskerQuantile}{the quantile to that whiskers of boxplots will extend, t.m. (1 - quantile*2) portion of data will be inside the whiskers}

\item{MinProteinAmount}{the minimum valid protein amount in the input data, protein amounts below this threshold are considered as NA}

\item{DropInvalidLogRatio}{if TRUE log-ratios outside validity range will be dropped}

\item{LogRatioValidityRange}{the log-ratio validity range}

\item{LogRatioPlotRange}{the log-ratio range for plotting}

\item{LogIntensityPlotRange}{the log2-intensity range for plotting}

\item{MaxLogRatioForAUQC}{the maximum value for AUQC quantification}

\item{Log2IntensityRanges}{the log2-intensity ranges for range based plots}

\item{CenterLogRatioByBackground}{should we center all log-ratios by the median log-ratio of the background species}

\item{NormalizeAmountsToPPM}{should we normalize amounts to ppms}

\item{PlotWidth}{the plot width in inches used in pdf files}

\item{PlotHeight}{the plot height in inches used in pdf files}

\item{PlotCurveLineWidth}{the line thickness for curves in plots}

\item{PlotLegendLineWidth}{the line thickness in legends}

\item{PlotPointSize}{the point size}

\item{ScatterPlotPointType}{the point type}

\item{PlotPointMinAlpha}{the minimum alpha value used for coloring points in sparse regions}

\item{AxisLabelSize}{the relative font size for axis labels}

\item{AxisAnnotationSize}{the relative font size for axis labels}

\item{AxisLineThickness}{the line thickness of axes}

\item{par}{the graphical parameters like mar, mgp, las, ... to set for the plot canvases.}
}
\description{
change some of the initialized LFQbench configuration settings.
changeConfig allows to change some configuration settings without loading default values.
}

