% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.com.R
\name{createNumericPairs}
\alias{createNumericPairs}
\title{createNumericPairs}
\usage{
createNumericPairs(from, to)
}
\arguments{
\item{from}{variable from the start number of the range}

\item{to}{variable end of the number range}
}
\description{
Create pairs for a range of integer numbers
}

