% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{emptyPlot}
\alias{emptyPlot}
\title{empyPlot}
\usage{
emptyPlot(xRange = c(0, 1), yRange = c(0, 1), lwd = 1, grid = T,
  showXlab = T, showYlab = T, axes = T, cex.axis = 1)
}
\arguments{
\item{xRange}{}

\item{yRange}{}

\item{lwd}{}

\item{grid}{}

\item{showXlab}{}

\item{showYlab}{}

\item{axes}{}

\item{cex.axis}{}
}
\description{
This function show empty plot with coordinate system in given ranges x/y-Axes are plotted on the bottom/left side with given line width grid can be switched off
}

