% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{evalCommandLineArguments}
\alias{evalCommandLineArguments}
\title{evalCommandLineArguments
This function evaluates the commandLine Arguments}
\usage{
evalCommandLineArguments()
}
\description{
evalCommandLineArguments
This function evaluates the commandLine Arguments
}

