% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{expandRange}
\alias{expandRange}
\title{expandRange
expand a range by a relative factor.}
\usage{
expandRange(r = c(0, 1), f = 0)
}
\arguments{
\item{r}{the range as a vector of two numbers in increasing order}

\item{f}{the ralative expansion factor, e.g. 
f=0 will not affect the original range,
f=0.2 will expand the range by 20%, t.m. 10% to each side}
}
\description{
expandRange
expand a range by a relative factor.
}
