% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{explainMetrics}
\alias{explainMetrics}
\title{explainMetrics}
\usage{
explainMetrics()
}
\description{
This funtions prints a short examplantion about the computed metrics for the analysis
}

