% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{fixFileExt}
\alias{fixFileExt}
\title{fixFileExt}
\usage{
fixFileExt(file, sep)
}
\arguments{
\item{file}{file name or path}

\item{sep}{column separator character}
}
\value{
fixed file name or path
}
\description{
fix the given file name by applying an appropriate file extension
that is guessed by column separator character provided
the new file extension will be
csv for ',' or ';'
tsv for '\t'
txt for anything else
}
