% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.cfg.r
\name{FSWE.addModification}
\alias{FSWE.addModification}
\title{FSWE.addModification}
\usage{
FSWE.addModification(modificationRegExps, UniModStrings)
}
\arguments{
\item{modificationRegExps}{Modification regular expressions as it may be found at the software tool report (input of FSWE). Use here the whole string you would like to substitute by a UniMod string}

\item{UniModStrings}{UniMod string that will substitute modificationString at FSWE outputs.}
}
\description{
Adds a new modification string to the modifications parser.
}
\examples{
FSWE.addModification( "\\\\[Oxi\\\\]", "\\\\(UniMod:35\\\\)")
}

