% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fswe.cfg.r
\name{FSWE.addSoftwareConfiguration}
\alias{FSWE.addSoftwareConfiguration}
\title{FSWE.addSoftwareConfiguration}
\usage{
FSWE.addSoftwareConfiguration(softwareName, input_format = "wide",
  input.extension = "*.csv$", nastrings = "NA", protein_input = F,
  filename.var = NA, quantitative.var = NA, quantitative.var.tag = NA,
  protein.var = NA, sequence.mod.var = NA, charge.var = NA,
  decoy.var = NA, decoy.tags = c("DECOY_", "reverse"), fdr.var.tag = NA,
  qvalue.var = NA, q_filter_threshold = NA, sheet.data = NA,
  sheet.fdr = NULL)
}
\arguments{
\item{softwareName}{:   a name for this configuration set.}

\item{input_format}{:   data format of the file. Options: "wide", "long"}

\item{input.extension}{:   file extension.
Warning: FSWE interprets file extensions to read the files.
Ensure that files named i.e. ".tsv" are actually tab separated
values files. Do not use the extension ".xls" in tab separated
values files.}

\item{nastrings}{:   string used at the file to label not available / not a number data.
Examples: "NA", "#N/A", "NaN".}

\item{protein_input}{:   Boolean. Set to TRUE if the files contain only protein information
(no sequence/precursors reports).}

\item{filename.var}{:   when using "long" report formats, column head name of the injection
file names.}

\item{quantitative.var}{:   column head name of the quantitative values.}

\item{quantitative.var.tag}{:   when using "wide" report formats, indicate a common tag for all
quantitative columns.}

\item{protein.var}{:   column head name of the protein names. It must contain a species tag.}

\item{sequence.mod.var}{:   column head name of the sequence (including modifications).}

\item{charge.var}{:   column head name of the precursor charge state.}

\item{decoy.var}{:   (optional) column head name of is_decoy (yes/no)}

\item{decoy.tags}{:   vector of tags for decoy proteins/peptides.
Example: c("DECOY_", "reverse")}

\item{fdr.var.tag}{:   when using excel formats (SWATH 2.0), indicate a tag to locate
the FDR columns.}

\item{qvalue.var}{:   (optional) If filtering by Q-value is required, column head name
of the Q-Value.}

\item{q_filter_threshold}{:   (optional) Q-value threshold, if you want to filter results by Q-value.
Note that you must specify then the qvalue.var (see above).}

\item{sheet.data}{:   when using excel formats, name of the sheet containing the
quantitative values.}

\item{sheet.fdr}{:   when using excel formats, name of the sheet containing the
                                 False Discovery Rate values.}
}
\description{
add a new software and its parameters to the list of softwares
}

