% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fswe.generate_reports.r
\name{FSWE.generateReports}
\alias{FSWE.generateReports}
\title{FSWE.generateReports}
\usage{
FSWE.generateReports(experimentFile, plotHistogram = F, plotHistNAs = F,
  reportSequences = F, peptide.list = NULL, protein.list = NULL,
  singleHits = F, softwareSource = "guess",
  working_dir = LFQbench.Config$DataRootFolder, keep_original_names = FALSE,
  outputFileNameSuffix = "")
}
\arguments{
\item{experimentFile}{}
}
\description{
FSWE.generateReports
}

