% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fswe.cfg.r
\name{FSWE.initConfiguration}
\alias{FSWE.initConfiguration}
\title{FSWE.initConfiguration}
\usage{
FSWE.initConfiguration(injectionNames, speciesTags)
}
\description{
define software list and the list of initialization function for each software
}

