% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.cfg.r
\name{FSWE.listModifications}
\alias{FSWE.listModifications}
\title{FSWE.listModifications}
\usage{
FSWE.listModifications()
}
\value{
list of modifications used by FSWE
}
\description{
FSWE.listModifications
}

