% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.scale_intensities.R
\name{FSWE.scaleIntensities}
\alias{FSWE.scaleIntensities}
\title{FSWE.scaleIntensities}
\usage{
FSWE.scaleIntensities(working_dir = LFQbench.Config$DataRootFolder,
  fileModelingPattern = "_peptides", filesToModelPattern = ".[\\\\.].",
  filesNotToModelPattern = "FILE_I_DONT_WANT_TO_HAVE",
  softwareReference = "PeakView", softToModelMap = list(),
  softwareLabels = NULL, outputFileNameSuffix = NULL)
}
\arguments{
\item{working_dir}{LFQbench-root directory with the subfolder input containing files processed by FSWE that you want to scale}

\item{fileModelingPattern}{pattern for files you want to use for modeling}

\item{filesToModelPattern}{pattern for files to apply the scaling models}

\item{softwareReference}{software name for scaling reference}
}
\description{
It scales FSWE-processed files to a common intensity scale defined by one software reference.
}
\details{
Assumption: since we lose  at this point the peak retention times, 
we will assume in this script that most of the peaks detected by different software tools match.
}

