% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fswe.cfg.r
\name{FSWE.setDatasetInjectionNames}
\alias{FSWE.setDatasetInjectionNames}
\title{FSWE.setDatasetInjectionNames}
\usage{
FSWE.setDatasetInjectionNames(theData)
}
\arguments{
\item{theData}{data.frame having}
}
\description{
set injections names used in the input data
}

