% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.cfg.r
\name{FSWE.setSpeciesTags}
\alias{FSWE.setSpeciesTags}
\title{FSWE.setSpeciesTags}
\usage{
FSWE.setSpeciesTags(speciesTags)
}
\arguments{
\item{speciesTags}{named list of tags. Names of elements must contain all LFQbench.Config$AllSpeciesNames}
}
\description{
set expressions for identifying species in protein names, e.g. _HUMAN, _YEAST, ...
}

