% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.com.r
\name{FSWE.simExperiment}
\alias{FSWE.simExperiment}
\title{FSWE.simExperiment}
\usage{
FSWE.simExperiment(numReplicates, species, speciesRatios, stdDevRatios,
  numProteinsSpecies, peptidesPerProtein, proteinAbundanceDistribution,
  stdDeviationFactorMS, BackgroundSignalLevel, NMARFactor, MARFactor,
  ProteinAbundanceErrorFactor, weibullShape = 0.5)
}
\arguments{
\item{numReplicates}{number of replicates of the experiment (same number for samples A and B)}

\item{species}{vector with species tag names. Example: species = c("HUMAN", "YEAST", "ECOLI")}

\item{speciesRatios}{vector with expected log ratios for each species. Example: speciesRatios = c(0.0, 1.0, -2.0)}

\item{stdDevRatios}{biological/prep variation. Example: stdDeviations = c(0.05, 0.1, 0.1)}

\item{numProteinsSpecies}{vector with number of proteins simulated per species. Example: numProteins = c(2000, 1500, 1000)}

\item{peptidesPerProtein}{vector with the average number of peptides per protein simulated. Example: peptidesPerProtein = c(10, 8, 5)}

\item{proteinAbundanceDistribution}{vector containing mean and sd values defining the protein distribution intensities desired. Example: proteinAbundanceDistribution = c(10.0, 5.0)}

\item{stdDeviationFactorMS}{standard deviation factor (0 to 1) due to MS. Example: stdDeviationFactorMS = 0.03}

\item{BackgroundSignalLevel}{A threshold value limiting the signal. Every peptide signal estimated under this will be transformed to NA.}

\item{NMARFactor}{factor of Not Missing At Random missing values in the experiment. They are intensity-dependant!}

\item{MARFactor}{factor of Missing At Random missing values in the experiment. Taken from a uniform distribution.}

\item{ProteinAbundanceErrorFactor}{error factor applied to the protein abundance.}

\item{weibullShape}{shape of the Weibull distribution used to model the probabilities peptides being selected as NMAR value.}
}
\description{
This function simulates a complete experiment that can be read by FSWE.
}

