% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fswe.cfg.r
\name{FSWE.switchSoftwareConfiguration}
\alias{FSWE.switchSoftwareConfiguration}
\title{FSWE.switchSoftwareConfiguration}
\usage{
FSWE.switchSoftwareConfiguration(softwareName)
}
\arguments{
\item{softwareName}{}
}
\description{
change the FSWE configuration to values for a specific software
}

