% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{getCombinedLogRatios}
\alias{getCombinedLogRatios}
\title{getCombinedLogRatios}
\usage{
getCombinedLogRatios(resultSets)
}
\arguments{
\item{resultSet}{}
}
\description{
This function compute the combined log ratios for the protein identifications and retrieve the values.
}

