% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{getCombinedProteinRSDs}
\alias{getCombinedProteinRSDs}
\title{getCombinedProteinRSDs}
\usage{
getCombinedProteinRSDs(ResultSets)
}
\arguments{
\item{ResultsSets}{}
}
\description{
This fucntion return the combined protein RSDs
}

