% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{getFileExt}
\alias{getFileExt}
\title{getFileExt}
\usage{
getFileExt(colSep)
}
\arguments{
\item{colSep}{column separator character}
}
\value{
a file extension csv/tsv/txt without '.'
}
\description{
generate appropriate file extension for a column separator: 
csv for ',' or ';'
tsv for '\t'
txt for anything else
}
