% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getOutlierIdx}
\alias{getOutlierIdx}
\title{getOutlierIdx
get outlier indices out of a range defined as mean +/- sdFac*SD}
\usage{
getOutlierIdx(flag, vals, flags, sdFac)
}
\arguments{
\item{flag}{identify values in @param flags by this value, e.g. a single species}

\item{vals}{vector of values, e.g. log-ratios}

\item{flags}{vector of flags, e.g. species}

\item{sdFac}{SD multiplier to define thresholds around the mean value where the outliers start}
}
\description{
getOutlierIdx
get outlier indices out of a range defined as mean +/- sdFac*SD
}

