% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{getQCFunction}
\alias{getQCFunction}
\title{getQCFunction}
\usage{
getQCFunction(ratios, ensureValueRange = c(0, 1))
}
\arguments{
\item{ensureValueRange}{ratio values range}

\item{rations}{the ratios to be exported}
}
\description{
This function calculate QC function
}

