% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getQuantileAccuracy}
\alias{getQuantileAccuracy}
\title{getQuantileAccuracy}
\usage{
getQuantileAccuracy(dataBySpecies, spcNames = LFQbench.Config$AllSpeciesNames,
  numberOfQuantiles = LFQbench.Config$NumberOfIntensityQuantiles)
}
\description{
This function compute the median deviation for log-ratios for each species in each range
}

