% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getQuantilePrecision}
\alias{getQuantilePrecision}
\title{getQuantilePrecision}
\usage{
getQuantilePrecision(dataBySpecies,
  spcNames = LFQbench.Config$AllSpeciesNames,
  numberOfQuantiles = LFQbench.Config$NumberOfIntensityQuantiles)
}
\description{
This function generate the standard deviation for log-ratios for each species in each range
}

