% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getQuantileSeparation}
\alias{getQuantileSeparation}
\title{getQuantileSeparation}
\usage{
getQuantileSeparation(dataBySpecies, spcNames,
  numberOfQuantiles = LFQbench.Config$NumberOfIntensityQuantiles)
}
\description{
This function calculates species separation ROC-AUC for a species pair
 in equally sized quantiles of intensity in first sample.
 
 @param dataBySpecies data structure as stored in ResultSet$result[[SamplePairIndex]]$data
 @param spcNames a vector of two species names
 @param numberOfQuantiles the number of quantiles
}

