% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.proc.R
\name{getRangedAccuracy}
\alias{getRangedAccuracy}
\title{getRangedAccuracy}
\usage{
getRangedAccuracy(dataBySpecies, ranges = cfg$Log2IntensityRanges,
  spcNames = cfg$AllSpeciesNames)
}
\arguments{
\item{dataBySpecies}{}

\item{ranges}{}

\item{spcNames}{}
}
\description{
This median deviation for log-ratios for each species in each range
}

