% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getRangedPrecision}
\alias{getRangedPrecision}
\title{getRangePrecision}
\usage{
getRangedPrecision(dataBySpecies,
  ranges = LFQbench.Config$Log2IntensityRanges,
  spcNames = LFQbench.Config$AllSpeciesNames)
}
\description{
This function generate the standard deviation for log-ratios for each species in each range
}

