% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getRangedSepRate}
\alias{getRangedSepRate}
\title{getRangedSepRate}
\usage{
getRangedSepRate(dataBySpecies, spcNames,
  ranges = LFQbench.Config$Log2IntensityRanges)
}
\arguments{
\item{spcNames}{}
}
\description{
This function calculate species separation ROC-AUC for a species pair
}

