% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getSepRate}
\alias{getSepRate}
\title{getSepRate}
\usage{
getSepRate(dataBySpecies, spcNames)
}
\arguments{
\item{dataBySpecies}{}

\item{spcNames}{}
}
\description{
This function calculate species separation ROC-AUC for a species pair
}

