% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fswe.com.r
\name{guessOrganism}
\alias{guessOrganism}
\title{guessOrganism}
\usage{
guessOrganism(proteinid, species)
}
\arguments{
\item{proteinid}{char variable containing the Swissprot header.}

\item{species}{vector defining the organisms that are part of the experiment. It is defined at fswe.datasets.R

An example of species vector:
species <- vector(mode="list", length=3)
names(species) <- c("HUMAN", "YEAST", "ECOLI")
species[[1]] <- "_HUMAN"
species[[2]] <- "_YEAS"
species[[3]] <- "_ECOLI"}
}
\description{
this function uses Swissprot annotations i.e. QXXXXX_HUMAN
to guess what's the species of the entry. The header (proteinid)
may have several entries, and thus it can belong to several
organisms. In that case, the function returns 'multiple'.
}

