% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{guessSep}
\alias{guessSep}
\title{guessSep
This function assigns a separator for separated values files readers
depending on file extension}
\usage{
guessSep(filename)
}
\arguments{
\item{filename}{}
}
\description{
guessSep
This function assigns a separator for separated values files readers
depending on file extension
}

