% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{guessSoftwareSource}
\alias{guessSoftwareSource}
\title{guessSoftwareSource}
\usage{
guessSoftwareSource(filename, software_sources)
}
\arguments{
\item{filename}{The file name should start by one of the software sources}

\item{software_sources}{a list defined at the variables file.}
}
\description{
guessSoftwareSource
}

