% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.batch.r
\name{LFQbench.batchProcessFiles}
\alias{LFQbench.batchProcessFiles}
\title{LFQbench.batchProcessFiles}
\usage{
LFQbench.batchProcessFiles(inputFiles, storeResultSetsImage = T,
  storeSampleMeans = F, storeSpeciesSeparation = F, storeIDs = F,
  storeIdStats = T, storeLogRatios = F, storePlots = T,
  storeSampleCVs = F, storeMetrics = T)
}
\arguments{
\item{inputFiles}{a list of files to process}
}
\description{
Wrapper function for processing a batch of input files.
Plots, metrics and log files will be automatically put into 
subfolders of a root folder.
PLEASE DEFINE THE ROOT FOLDER BEFORE RUNNING THIS FUNCTION!!!
Expected structure of root folder:
}
\details{
- subfolder "input" contains automatically discovered input files
 - subfolder "log" contains calculated metrics, stored result sets and other exported information.
 - subfolder "plot" contains generated plots.

In case the given root folder is empty, all needed file structures will be created automatically.
You can use the freshly created structures to put your input files and rerun LFQbench.
}

