% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{LFQbench.explainMetrics}
\alias{LFQbench.explainMetrics}
\title{LFQbench.explainMetrics}
\usage{
LFQbench.explainMetrics()
}
\description{
This function verbose the metrics of the experiment
}

