% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{LFQbench.getMetrics}
\alias{LFQbench.getMetrics}
\title{LFQbench.getMetrics}
\usage{
LFQbench.getMetrics(resultSet)
}
\arguments{
\item{resultSet}{}
}
\description{
calculate metrics for a result set
}

