% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.cfg.r
\name{LFQbench.initConfiguration}
\alias{LFQbench.initConfiguration}
\title{LFQbench.initConfiguration}
\usage{
LFQbench.initConfiguration(DataRootFolder = getwd(),
  SampleComposition = data.frame(species = c("HUMAN", "YEAST", "ECOLI"), A =
  c(65, 30, 5), B = c(65, 15, 20)), BackgroundSpeciesName = "HUMAN",
  InputExtensionPattern = "\\\\..sv$", CsvColumnSeparator = "\\t",
  CsvDecimalPointChar = ".", BoxPlotWhiskerQuantile = 0.025,
  MinProteinAmount = 1e-06, DropInvalidLogRatio = T,
  LogRatioValidityRangeSDFactor = 5, LogRatioPlotRange = c(-4, 4),
  LogIntensityPlotRange = NULL, MaxLogRatioForAUQC = 2,
  NumberOfIntensityQuantiles = 3, CenterLogRatioByBackground = T,
  NormalizeAmountsToPPM = F, PlotWidth = 6, PlotHeight = 5,
  PlotCurveLineWidth = 2, PlotLegendLineWidth = 4, PlotPointSize = 1.5,
  ScatterPlotPointType = 20, ScatterPlotPointMinAlpha = 0.5,
  ScatterPlotLowessBandWidth = 0.8, AxisLabelSize = 2,
  AxisAnnotationSize = 2, AxisLineThickness = 2, AxisXLabelNumDiv = NULL,
  AxisYLabelNumDiv = NULL, par = list(mar = c(4.5, 6, 0.5, 0.5), mgp =
  c(3.7, 1.5, 0), las = 1))
}
\arguments{
\item{DataRootFolder}{the parent folder for batch processing}

\item{SampleComposition}{the quantitative composition of hybrid proteome samples}

\item{BackgroundSpeciesName}{the name of non regulated background species}

\item{InputExtensionPattern}{the pattern of file extension for batch input files}

\item{CsvColumnSeparator}{the input file column delimiter character}

\item{CsvDecimalPointChar}{the input file decimal point character}

\item{BoxPlotWhiskerQuantile}{the quantile to that whiskers of boxplots will extend, t.m. (1 - quantile*2) portion of data will be inside the whiskers}

\item{MinProteinAmount}{the minimum valid protein amount in the input data, protein amounts below this threshold are considered as NA}

\item{DropInvalidLogRatio}{if TRUE log-ratios outside validity range will be dropped}

\item{LogRatioValidityRangeSDFactor}{validity range is calculated as mean +/- SD * SDFactor (separately for each species)}

\item{LogRatioPlotRange}{the log-ratio range for plotting}

\item{LogIntensityPlotRange}{the log2-intensity range for plotting}

\item{MaxLogRatioForAUQC}{the maximum value for AUQC quantification}

\item{NumberOfIntensityQuantiles}{the number of parts for splitting the data for metrics calculation}

\item{CenterLogRatioByBackground}{should we center all log-ratios by the median log-ratio of the background species}

\item{NormalizeAmountsToPPM}{should we normalize amounts to ppms}

\item{PlotWidth}{the plot width in inches used in pdf files}

\item{PlotHeight}{the plot height in inches used in pdf files}

\item{PlotCurveLineWidth}{the line thickness for curves in plots}

\item{PlotLegendLineWidth}{the line thickness in legends}

\item{PlotPointSize}{the point size}

\item{ScatterPlotPointType}{the point type}

\item{ScatterPlotPointMinAlpha}{the minimum alpha value used for coloring points in sparse regions}

\item{ScatterPlotLowessBandWidth}{the lowess locality parameter f, use values (local) 0 < f <= 1 (global)}

\item{AxisLabelSize}{the relative font size for axis labels}

\item{AxisAnnotationSize}{the relative font size for axis labels}

\item{AxisLineThickness}{the line thickness of axes}

\item{AxisXLabelNumDiv}{divide the number of labels displayed in axis X}

\item{AxisYLabelNumDiv}{divide the number of labels displayed in axis Y}

\item{par}{the graphical parameters like mar, mgp, las, ... to set for the plot canvases.}
}
\description{
Initialization of LFQbench configuration settings.
Executing the function without parameters will use default values.
Use function LFQbench.changeConfiguration to change settings after initialization.
}

