% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{LFQbench.plotSampleComposition}
\alias{LFQbench.plotSampleComposition}
\title{LFQbench.plotSampleComposition}
\usage{

  LFQbench.plotSampleComposition(sampleAmounts = data.frame(LFQbench.Config$SampleComposition,
  row.names = 1), bgColors = NULL, fgColors = NULL, labMainSize = 1.5,
  pdfFile = paste(LFQbench.Config$PlotFilesLocation, "/SampleComposition.pdf",
  sep = ""), pdfWidth = 2, pdfHeight = 3, pdfFontSize = 9,
  pdfFontFamily = "Helvetica", pdfMar = c(2.3, 2.8, 0.5, 4.6))
}
\description{
draw the sample composition as a spine plot
}

