% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{LFQbench.plotSpeciesLegends}
\alias{LFQbench.plotSpeciesLegends}
\title{LFQbench.plotSpeciesLegends}
\usage{
LFQbench.plotSpeciesLegends(plotFolder = LFQbench.Config$PlotFilesLocation)
}
\description{
plot vertical and horizontal species legends
 @param plotFolder the folder, where plots should go to
 @export
}

