% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.cfg.r
\name{LFQbench.processConfig}
\alias{LFQbench.processConfig}
\title{recalculate some parameters from user defined configuration}
\usage{
LFQbench.processConfig()
}
\description{
recalculate some parameters from user defined configuration
}

