% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{LFQbench.processDocSet}
\alias{LFQbench.processDocSet}
\title{LFQbench.processDocSet}
\usage{
LFQbench.processDocSet(DocSet)
}
\arguments{
\item{DocSet}{is the aata to be processed}
}
\description{
This function process the data before generate the charts
}

