% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.cfg.r
\name{LFQbench.setDataRootFolder}
\alias{LFQbench.setDataRootFolder}
\title{LFQbench.setDataRootFolder}
\usage{

  LFQbench.setDataRootFolder(rootFolder = ifelse(file.exists(LFQbench.Config$DataRootFolder),
  LFQbench.Config$DataRootFolder, getwd()), createSubfolders = T)
}
\arguments{
\item{rootFolder}{the path to the root folder}
}
\description{
change the location of root folder for batch processing
}

