% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{LFQbench.showAllSpeciesQC}
\alias{LFQbench.showAllSpeciesQC}
\title{LFQbench.showAllSpeciesQC}
\usage{
LFQbench.showAllSpeciesQC(samplePair)
}
\arguments{
\item{samplePair}{the result set data of a hybrid proteome sample pair}
}
\description{
draw quantification curves
}

