% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{LFQbench.showMetrics}
\alias{LFQbench.showMetrics}
\title{LFQbench.showMetrics}
\usage{
LFQbench.showMetrics(m)
}
\arguments{
\item{m}{the metrics object}
}
\description{
verbose the metrics of the experiment
}

