% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{LFQbench.showScatterAndDensityPlot}
\alias{LFQbench.showScatterAndDensityPlot}
\title{LFQbench.showScatterAndDensityPlot}
\usage{
LFQbench.showScatterAndDensityPlot(samplePair, showLegend = F,
  showRegLines = F, scatterPlotWidth = 0.8, at.x = NULL, at.y = NULL)
}
\arguments{
\item{samplePair}{the result set data of a hybrid proteome sample pair}

\item{showLegend}{display the legend}

\item{showRegLines}{display regression curves}

\item{scatterPlotWidth}{the window portion used for scatter plot}
}
\description{
draw scatter plot with an attached density plot
}

