% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.config.R
\name{loadConfig}
\alias{loadConfig}
\title{loadConfig}
\usage{
loadConfig(dataFolder)
}
\arguments{
\item{String}{Data folder where the original data for the anlysis will be store and handle}
}
\description{
this function configure the enviroment for the main functions and functionalities of the project
}

