% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.com.R
\name{loadLibrary}
\alias{loadLibrary}
\title{Load and install function}
\usage{
loadLibrary(x)
}
\arguments{
\item{x}{A name of a library to be used or uploaded}
}
\value{
This funtion only install the dependecies libraries
}
\description{
Load function upload different libraries and dependencies in the package
}

