% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{logIdStatistics}
\alias{logIdStatistics}
\title{logIdStatistics}
\usage{
logIdStatistics(resultSets)
}
\arguments{
\item{resultSets}{}
}
\description{
This function trace the Statistics related with each ID
}

