% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{makeDocSet}
\alias{makeDocSet}
\title{makeDocSet}
\usage{
makeDocSet(inFile, inputPath = cfg$InputFilesLocation,
  plotPath = cfg$PlotFilesLocation, logPath = cfg$LogFilesLocation)
}
\arguments{
\item{inFile}{}
}
\description{
construct a document set for a file by paths from user configuration
}

