% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{makeEverythingInOneFolderDocSet}
\alias{makeEverythingInOneFolderDocSet}
\title{makeEverythingInOneFolderDocSet}
\usage{
makeEverythingInOneFolderDocSet(inFile)
}
\description{
construct a doc set using the input file's path
}

