% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{makeScatter}
\alias{makeScatter}
\title{makeScatter}
\usage{
makeScatter(samplePair, showLegend = F, showRegLines = F,
  showExpLines = T, useCfgColor = T)
}
\arguments{
\item{samplePair}{the result set data of a sample pair}

\item{showLegend}{display the legend}

\item{showRegLines}{display regression curves}

\item{showExpLines}{display expectation lines}

\item{useCfgColor}{ignore result set colors and use colors from current configuration settings}
}
\description{
make a scatter plot with smoothed colors
}

