% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.com.r
\name{median_all}
\alias{median_all}
\title{median_all}
\usage{
median_all(values, minimum = 1)
}
\arguments{
\item{values}{the vector with the values.}

\item{minimum}{minimum number of valid (not NA) values allowed in order to return a value.}
}
\description{
median_all
}

