% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.com.r
\name{median_quench}
\alias{median_quench}
\title{median_quench}
\usage{
median_quench(values, minimum = 1, minPepsToRemove = 10, removeTop = 3)
}
\arguments{
\item{values}{the vector with the values.}

\item{minimum}{minimum number of valid (not NA) values allowed in order to return a value.}

\item{minPepsToRemove}{minimum number of peptides the protein must have in order to remove the top N peptides}

\item{removeTop}{number of peptides to be removed when the protein has more than minPepsToRemove}
}
\description{
median_quench
}

