% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{mkdir}
\alias{mkdir}
\title{mkdir
This recursively create a directory if it does not exist}
\usage{
mkdir(dirName)
}
\arguments{
\item{dirName}{the directory to create the directory}
}
\description{
mkdir
This recursively create a directory if it does not exist
}

