% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{moveFiles}
\alias{moveFiles}
\title{moveFiles}
\usage{
moveFiles(srcDir, tarDir, namePattern = "\\\\..+", rmSrc = F)
}
\arguments{
\item{srcDir}{the origin folder}

\item{tarDir}{the target folder}

\item{namePattern}{the file name pattern for files to be moved}

\item{rmSrc}{if the origin folder should be removed, it can only be removed if it is empty}
}
\description{
move all files (with any file extension) to a new directory
}

