% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{plotLogRatios}
\alias{plotLogRatios}
\title{plotLogRatios}
\usage{
plotLogRatios(logRatios, File)
}
\arguments{
\item{logRatios}{log ratio values}

\item{file}{}
}
\description{
This funtion plot the Log Rations in a File
}

