% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{plotProteinDispersionBySample}
\alias{plotProteinDispersionBySample}
\title{plotProteinDispersionBySample}
\usage{
plotProteinDispersionBySample(d)
}
\arguments{
\item{d}{}
}
\description{
This draw protein dispersion between replicates by sample
}

