% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{plotResultSet}
\alias{plotResultSet}
\title{plotResultSet}
\usage{
plotResultSet(resSet)
}
\arguments{
\item{resSet}{result set from the analylis}
}
\description{
The resultset fucntion adjust the values for the pdf generation
}

