% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{plotSpeciesLegend}
\alias{plotSpeciesLegend}
\title{plotSpeciesLegend}
\usage{
plotSpeciesLegend(pos = "top", ...)
}
\arguments{
\item{pos}{position of the legend}
}
\description{
This function plot the species Legend in the final chart
}

