% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{plotToFile}
\alias{plotToFile}
\title{plotToFile}
\usage{
plotToFile(file, plotFunc, ...)
}
\arguments{
\item{file}{the pdfFile}

\item{plotFunc}{the plot function}
}
\description{
Plot user content to file by using LFQbench canvas configuration.
}

