% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.com.R
\name{+}
\alias{+}
\title{+ function}
\usage{
"+"(...)
}
\arguments{
\item{x}{set of String values to be concatenated}
}
\value{
a resulted String
}
\description{
Define + for String concatenation
}

