% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.cfg.r
\name{processConfig}
\alias{processConfig}
\title{recalculate some parameters from user defined configuration}
\usage{
processConfig()
}
\description{
recalculate some parameters from user defined configuration
}

