% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{processData}
\alias{processData}
\title{processData}
\usage{
processData(DocSet)
}
\arguments{
\item{DocSet}{is the Data to be processed}
}
\description{
This function process the Data before generate the charts
}

