% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{processDocSet}
\alias{processDocSet}
\title{processDocSet}
\usage{
processDocSet(DocSet)
}
\arguments{
\item{DocSet}{is the aata to be processed}
}
\description{
This function process the data before generate the charts
}

