% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{processFile}
\alias{processFile}
\title{processFile}
\usage{
processFile(file)
}
\arguments{
\item{file}{is the data to be processed}
}
\description{
This function process the data and generates a result set
}

