% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{saveLogRatios}
\alias{saveLogRatios}
\title{saveLogRatios}
\usage{
saveLogRatios(rs)
}
\arguments{
\item{resultSet}{}
}
\description{
This function sabe the logRatios of the protein identifications
}

