% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{saveSampleMeans}
\alias{saveSampleMeans}
\title{saveSampleMeans}
\usage{
saveSampleMeans(resultSet)
}
\arguments{
\item{resultSet}{}
}
\description{
This function save the mean for every Sample
}

