% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{saveSampleRSD}
\alias{saveSampleRSD}
\title{saveSampleRSD}
\usage{
saveSampleRSD(resultSet)
}
\arguments{
\item{resultSet}{}
}
\description{
This function store the variation coeeficients for each sample
}

