% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{saveSpeciesSeparation}
\alias{saveSpeciesSeparation}
\title{saveSpeciesSeparation}
\usage{
saveSpeciesSeparation(resultSet)
}
\arguments{
\item{resultSet}{}
}
\description{
This function store the species for each protein identification
}

