% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{scaleColor}
\alias{scaleColor}
\title{scaleColor}
\usage{
scaleColor(col, scale = 1.5, alpha = 1)
}
\arguments{
\item{col}{Color}

\item{scale}{}

\item{alpha}{}
}
\description{
This function make a color darker or lighter by scaling
}

