% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.cfg.r
\name{setRootFolder}
\alias{setRootFolder}
\title{setRootFolder}
\usage{
setRootFolder(rootFolder = ifelse(file.exists(cfg$DataRootFolder),
  cfg$DataRootFolder, getwd()), createSubfolders = T)
}
\arguments{
\item{rootFolder}{the path to the root folder}
}
\description{
change the location of root folder for batch processing
}

