% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{showAllSpeciesQC}
\alias{showAllSpeciesQC}
\title{showAllSpeciesQC}
\usage{
showAllSpeciesQC(samplePair)
}
\arguments{
\item{samplePair}{}
}
\description{
This fuction draw quantification curve
}

