% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{showDistributionDensityPlot}
\alias{showDistributionDensityPlot}
\title{showDistributionDensityPlot}
\usage{
showDistributionDensityPlot(samplePair, showLegend = F)
}
\arguments{
\item{samplePair}{the result set data of a hybrid proteome sample pair}

\item{showLegend}{display the legend}
}
\description{
draw log-ratio distribution kernel density plot
}

