% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{showLogRatioBoxPlot}
\alias{showLogRatioBoxPlot}
\title{showLogRatioBoxPlot}
\usage{
showLogRatioBoxPlot(samplePair)
}
\arguments{
\item{samplePair}{}
}
\description{
This function draw log ratio quiantile based boxplot
}

