% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{showQuantBarPlot}
\alias{showQuantBarPlot}
\title{showQuantBarPlot}
\usage{
showQuantBarPlot(samplePair)
}
\arguments{
\item{samplePair}{the result set data of a hybrid proteome sample pair}
}
\description{
draw barplot showing quantification bars
}

