% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{showScatterAndBoxPlot}
\alias{showScatterAndBoxPlot}
\title{showScatterAndBoxPlot}
\usage{
showScatterAndBoxPlot(samplePair, showLegend = F, showRegLines = F)
}
\arguments{
\item{samplePair}{}

\item{showLegend}{}

\item{showRegLines}{}
}
\description{
This function draw scatter- and boxplot
}

