% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{showScatterAndDensityPlot}
\alias{showScatterAndDensityPlot}
\title{showScatterAndDensityPlot}
\usage{
showScatterAndDensityPlot(samplePair, showLegend = F, showRegLines = F)
}
\arguments{
\item{samplePair}{}

\item{showRegLines}{}

\item{showLengend}{}
}
\description{
This fucntion draw scatter- and density plot
}

