% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{showScatterPlot}
\alias{showScatterPlot}
\title{showScatterPlot}
\usage{
showScatterPlot(samplePair, showLegend = F, showRegLines = F)
}
\arguments{
\item{samplePair}{}

\item{showLegend}{}

\item{showRegLines}{}
}
\description{
This fucntion draw scatterplot
}

