% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{showSingleSpeciesQC}
\alias{showSingleSpeciesQC}
\title{showSingleSpeciesQC}
\usage{
showSingleSpeciesQC(resultSet, speciesName)
}
\arguments{
\item{resultSet}{the result set data}

\item{speciesName}{the name of a species}
}
\description{
draw quantification curves
}

