% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fswe.com.r
\name{single_hits}
\alias{single_hits}
\title{single_hits}
\usage{
single_hits(values)
}
\arguments{
\item{values}{}
}
\description{
This function filters vectors by returning NA if the vector has more than one valid element.
This is useful for summary functions in dplyr, for example
}

