% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.com.r
\name{take1stentry}
\alias{take1stentry}
\title{take1stentry}
\usage{
take1stentry(entries)
}
\arguments{
\item{entries}{char variable with the swissprot header}
}
\description{
this function takes the first entry of a swissprot database header
}

