% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{addAxes}
\alias{addAxes}
\title{addAxes}
\usage{
addAxes(lwd = 1, showXlab = T, showYlab = T, showXAxis = T,
  showYAxis = T, cex.axis = 1)
}
\arguments{
\item{lwd=1}{}

\item{showXlab=T}{}

\item{showYlab=T}{}

\item{showXAxis=T}{}

\item{showYAxis=T}{}

\item{cex.axis=1}{}
}
\description{
add normal axes having full lengths of plotting area
}

