% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{addAxes}
\alias{addAxes}
\title{addAxes}
\usage{
addAxes(lwd = 1, showXlab = T, showYlab = T, showXAxis = T,
  showYAxis = T, cex.axis = 1, at.x = NULL, at.y = NULL)
}
\arguments{
\item{at.x=NULL}{}
}
\description{
add normal axes having full lengths of plotting area
}

