% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{addXLab}
\alias{addXLab}
\title{addXLab
add an x-axis label shifted by marginShift from par()$mgp[1]}
\usage{
addXLab(xlab = "", marginShift = -0.2,
  cex.lab = LFQbench.Config$AxisLabelSize, ...)
}
\arguments{
\item{...}{other parameter passed to title(...)}
}
\description{
addXLab
add an x-axis label shifted by marginShift from par()$mgp[1]
}

