% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{createNumericPairs}
\alias{createNumericPairs}
\title{createNumericPairs}
\usage{
createNumericPairs(from, to)
}
\arguments{
\item{from}{}

\item{to}{}
}
\description{
This function creates pairs for a range of integer numbers (e.g. for from=1, to=4 the result is rbind(c(1,2),c(1,3),c(1,4),c(2,3),c(2,4),c(3,4))
}

