% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{evalCommandLineArguments}
\alias{evalCommandLineArguments}
\title{evalCommandLineArguments
This function eval the commandLine Arguments}
\usage{
evalCommandLineArguments()
}
\description{
evalCommandLineArguments
This function eval the commandLine Arguments
}

