% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fswe.cfg.r
\name{FSWE.addSoftwareConfiguration}
\alias{FSWE.addSoftwareConfiguration}
\title{FSWE.addSoftwareConfiguration}
\usage{
FSWE.addSoftwareConfiguration(softwareName, quantitative.var = NA,
  quantitative.var.tag = NA, protein.var = NA, filename.var = NA,
  sequence.mod.var = NA, charge.var = NA, q_filter_threshold = NA,
  decoy.var = NA, sheet.fdr = NULL, protein_input = F,
  input.extension = "*.csv$", nastrings = "NA", decoy.tags = c("DECOY_",
  "reverse"), fdr.var.tag = NA, input_format = "wide", qvalue.var = NA,
  sheet.data = NA)
}
\arguments{
\item{softwareName}{}

\item{quantitative.var}{= NA}

\item{quantitative.var.tag}{= NA}

\item{protein.var}{= NA}

\item{filename.var}{= NA}

\item{sequence.mod.var}{= NA}

\item{charge.var}{= NA}

\item{q_filter_threshold}{= NA}

\item{decoy.var}{= NA}

\item{sheet.fdr}{= NULL}

\item{protein_input}{= F}

\item{input.extension}{= "*.csv$"}

\item{nastrings}{= "NA"}

\item{decoy.tags}{= c("DECOY_""reverse")}

\item{fdr.var.tag}{= NA}

\item{input_format}{= "wide"}

\item{qvalue.var}{= NA}

\item{sheet.data}{= NA}
}
\description{
add a new software and its parameters to the list of softwares
}

