% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.generate_reports.r
\name{FSWE.generateReports}
\alias{FSWE.generateReports}
\title{FSWE.generateReports}
\usage{
FSWE.generateReports(experimentFile, plotHistogram = F, plotHistNAs = F,
  reportSequences = F, peptide.list = NULL, protein.list = NULL,
  singleHits = F, softwareSource = "guess",
  working_dir = LFQbench.Config$DataRootFolder, keep_original_names = FALSE,
  outputFileNameSuffix = NULL, minimumHitsPerSample = 2)
}
\description{
It generates peptide and protein reports compatible with LFQbench, by following the data processing
suggested in (#Ref_to_article).
}

