% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fswe.com.r
\name{FSWE.simExperiment}
\alias{FSWE.simExperiment}
\title{FSWE.simExperiment}
\usage{
FSWE.simExperiment(numReplicates, species, speciesRatios, stdDevRatios,
  numProteinsSpecies, peptidesPerProtein, intDistribution, stdDeviationFactorMS,
  BackgroundSignalLevel, MissingValuesFactor, SignalNoiseFactor)
}
\arguments{
\item{numReplicates}{number of replicates of the experiment (same number for samples A and B)}

\item{species}{vector with species tag names. Example: species = c("HUMAN", "YEAST", "ECOLI")}

\item{speciesRatios}{vector with expected log ratios for each species. Example: speciesRatios = c(0.0, 1.0, -2.0)}

\item{stdDevRatios}{biological/prep variation. Example: stdDeviations = c(0.05, 0.1, 0.1)}

\item{numProteinsSpecies}{vector with number of proteins simulated per species. Example: numProteins = c(2000, 1500, 1000)}

\item{peptidesPerProtein}{vector with the average number of peptides per protein simulated. Example: peptidesPerProtein = c(10, 8, 5)}

\item{intDistribution}{vector containing mean and sd values defining the protein distribution intensities desired. Example: intDistribution = c(10.0, 5.0)}

\item{stdDeviationFactorMS}{standard deviation factor (0 to 1) due to MS. Example: stdDeviationFactorMS = 0.03}

\item{MissingValuesFactor}{factor of missing values in the experiment. They are intensity-dependant.}
}
\description{
This function simulates a complete experiment that can be read by FSWE.
}

