% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{getCombinedLogRatios}
\alias{getCombinedLogRatios}
\title{getCombinedLogRatios}
\usage{
getCombinedLogRatios(ResultSets)
}
\arguments{
\item{ResultSets}{}
}
\description{
This function get the Combined Log Ratios
}

