% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{getCombinedProteinRSDs}
\alias{getCombinedProteinRSDs}
\title{getCombinedProteinRSDs}
\usage{
getCombinedProteinRSDs(resultSets)
}
\arguments{
\item{resultSEt}{}
}
\description{
This function combined the protein variance cooeficents
}

