% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getQuantileAccuracy}
\alias{getQuantileAccuracy}
\title{getQuantileAccuracy}
\usage{
getQuantileAccuracy(dataBySpecies, spcNames = cfg$AllSpeciesNames,
  numberOfQuantiles = cfg$NumberOfIntensityQuantiles)
}
\arguments{
\item{dataBySpecies}{}

\item{spcNames}{}

\item{numberOfQuantiles}{}
}
\description{
This function compute the median deviation for log-ratios for each species in each range
}

