% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getQuantilePrecision}
\alias{getQuantilePrecision}
\title{getQuantilePrecision}
\usage{
getQuantilePrecision(dataBySpecies, spcNames = cfg$AllSpeciesNames,
  numberOfQuantiles = cfg$NumberOfIntensityQuantiles)
}
\arguments{
\item{dataBySpecies}{}

\item{spcNames}{}

\item{numberOfQuantiles}{}
}
\description{
This function generate the standard deviation for log-ratios for each species in each range
}

