% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getQuantileSeparation}
\alias{getQuantileSeparation}
\title{getQuantileSeparation}
\usage{
getQuantileSeparation(dataBySpecies, spcNames,
  numberOfQuantiles = cfg$NumberOfIntensityQuantiles)
}
\arguments{
\item{dataBySpecies}{data structure as stored in ResultSet$result[[SamplePairIndex]]$data}

\item{spcNames}{a vector of two species names}

\item{numberOfQuantiles}{the number of quantiles}
}
\description{
This function calculates species separation ROC-AUC for a species pair
 in equally sized quantiles of intensity in first sample.
}

