% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getRangedAccuracy}
\alias{getRangedAccuracy}
\title{getRangeAccuracy}
\usage{
getRangedAccuracy(dataBySpecies, ranges = Log2IntensityRanges,
  spcNames = LFQbench.Config$AllSpeciesNames)
}
\arguments{
\item{ranges}{}
}
\description{
This function compute the median deviation for log-ratios for each species in each range
}

