% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getRangedPrecision}
\alias{getRangedPrecision}
\title{getRangePrecision}
\usage{
getRangedPrecision(dataBySpecies, ranges = cfg$Log2IntensityRanges,
  spcNames = cfg$AllSpeciesNames)
}
\arguments{
\item{dataBySpecies}{}

\item{ranges}{}

\item{spcNames}{}
}
\description{
This function generate the standard deviation for log-ratios for each species in each range
}

