% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.proc.r
\name{getRangedSepRate}
\alias{getRangedSepRate}
\title{getRangedSepRate}
\usage{
getRangedSepRate(dataBySpecies, spcNames, ranges = cfg$Log2IntensityRanges)
}
\arguments{
\item{dataBySpecies}{}

\item{spcNames}{}

\item{ranges}{}
}
\description{
This function calculate species separation ROC-AUC for a species pair
}

