% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{LFQbench.plotSpeciesLegends}
\alias{LFQbench.plotSpeciesLegends}
\title{LFQbench.plotSpeciesLegends}
\usage{
LFQbench.plotSpeciesLegends()
}
\description{
plot vertical and horizontal species legends
}

