% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{LFQbench.plotToFile}
\alias{LFQbench.plotToFile}
\title{LFQbench.plotToFile}
\usage{
LFQbench.plotToFile(file, plotFunc, ...)
}
\description{
Plot user content to file by using LFQbench canvas configuration.
 
 @param file the pdfFile
 @param plotFunc the plot function
 @export
}

