% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{LFQbench.showScatterPlot}
\alias{LFQbench.showScatterPlot}
\title{LFQbench.showScatterPlot}
\usage{
LFQbench.showScatterPlot(samplePair, showLegend = F, showRegLines = F)
}
\arguments{
\item{samplePair}{the result set data of a hybrid proteome sample pair}

\item{showLegend}{display the legend}

\item{showRegLines}{display regression curves}
}
\description{
draw scatterplot
}

