% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{loadLibrary}
\alias{loadLibrary}
\title{loadLibrary
install and load a library}
\usage{
loadLibrary(x)
}
\arguments{
\item{x}{the name of the library to load}
}
\description{
loadLibrary
install and load a library
}

