% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{loadLibrary}
\alias{loadLibrary}
\title{loadLibrary
This function load an specific library for the package.}
\usage{
loadLibrary(x)
}
\arguments{
\item{x}{The library to be load}
}
\description{
loadLibrary
This function load an specific library for the package.
}

