% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{logIdStatistics}
\alias{logIdStatistics}
\title{logIdStatistics}
\usage{
logIdStatistics(resultSets)
}
\arguments{
\item{resultSets}{}
}
\description{
This function store the log statistics for the protein identifications
}

