% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{logIdStatistics}
\alias{logIdStatistics}
\title{logIdStatistics}
\usage{
logIdStatistics(resultSets, csvFile = paste(LFQbench.Config$LogFilesLocation,
  "/identification_statistics.csv", sep = ""))
}
\arguments{
\item{resultSets}{the result sets}

\item{csvFile}{the delimiter separated output file}
}
\description{
store identification statistics from a list of result sets
}

