% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{makeDocSet}
\alias{makeDocSet}
\title{makeDocSet}
\usage{
makeDocSet(inFile, inputPath = LFQbench.Config$InputFilesLocation,
  plotPath = LFQbench.Config$PlotFilesLocation,
  logPath = LFQbench.Config$LogFilesLocation)
}
\description{
construct a document set for a file by paths from user configuration
}

