% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{makeEverythingInOneFolderDocSet}
\alias{makeEverythingInOneFolderDocSet}
\title{makeEverythingInOneFolderDocSet}
\usage{
makeEverythingInOneFolderDocSet(inFile)
}
\arguments{
\item{inFile}{}
}
\description{
construct a doc set using the input file's path
}

