% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{makeScatter}
\alias{makeScatter}
\title{makeScatter
This make a simple scatter plot without globally changing graphical parameters}
\usage{
makeScatter(samplePair, showLegend = F, showRegLines = F)
}
\arguments{
\item{samplePair}{}

\item{showLegend}{}

\item{showRegLines}{}
}
\description{
makeScatter
This make a simple scatter plot without globally changing graphical parameters
}

