% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{mkdir}
\alias{mkdir}
\title{mkdir}
\usage{
mkdir(dirName)
}
\arguments{
\item{dirName}{the directory to create the directory}
}
\description{
This recursively create a directory if it does not exist
}

