% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.com.R
\name{mkdir}
\alias{mkdir}
\title{mkdir to create folders}
\usage{
mkdir(dirName)
}
\arguments{
\item{dirName}{the directory to be created}
}
\value{
do not return any information, only create the folder.
}
\description{
Create a folder recursivelly if it does not exist
}

