% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{moveFiles}
\alias{moveFiles}
\title{moveFiles}
\usage{
moveFiles(srcDir, tarDir, namePattern = "\\\\..+", rmSrc = F)
}
\description{
move all files (with any file extension) to a new directory
 @param srcDir the origin folder
 @param tarDir the target folder
 @param namePattern the file name pattern for files to be moved
 @param rmSrc if the origin folder should be removed, it can only be removed if it is empty
 @export
}

