% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{plotProteinDispersionBySample}
\alias{plotProteinDispersionBySample}
\title{plotProteinDispersionBySample}
\usage{
plotProteinDispersionBySample(d)
}
\arguments{
\item{d}{the data of a result set (resultSet$data)}
}
\description{
draw protein dispersion as cv between replicates by sample
}

