% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{plotResultSet}
\alias{plotResultSet}
\title{plotResultSet}
\usage{
plotResultSet(resSet, showScatterPlot = T, showScatterAndDensityPlot = T,
  showScatterAndBoxPlot = T, showBoxPlot = T, showDensityPlot = T)
}
\arguments{
\item{resSet}{a result set}

\item{showScatterPlot}{}

\item{showScatterAndDensityPlot}{}

\item{showScatterAndBoxPlot}{}

\item{showBoxPlot}{}

\item{showDensityPlot}{}
}
\description{
Function to generate the scatter plot
}

