% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{plotSampleComposition}
\alias{plotSampleComposition}
\title{plotSampleComposition}
\usage{
plotSampleComposition(sampleAmounts = data.frame(cfg$SampleComposition,
  row.names = 1), bgColors = NULL, fgColors = NULL, labMainSize = 1.5,
  pdfFile = paste(cfg$PlotFilesLocation, "/SampleComposition.pdf", sep = ""),
  pdfWidth = 2, pdfHeight = 3, pdfFontSize = 9,
  pdfFontFamily = "Helvetica", pdfMar = c(2.3, 2.8, 0.5, 4.6))
}
\arguments{
\item{sampleAmounts}{}

\item{bgColors}{}

\item{fgColors}{}

\item{labMainSize}{}

\item{pdfFile}{}

\item{pdfWidth}{}

\item{pdfHeight}{}

\item{pdfFontSize}{}

\item{pdfFontFamily}{}

\item{pdfMar}{}
}
\description{
draw the sample composition as a spine plot
}

