% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{plotSpeciesLegend}
\alias{plotSpeciesLegend}
\title{plotSpeciesLegend}
\usage{
plotSpeciesLegend(pos = "top", ...)
}
\arguments{
\item{pos}{the legend position on plot area}
}
\description{
This function add the species legend to a plot
}

