% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{plotSpeciesLegend}
\alias{plotSpeciesLegend}
\title{plotSpeciesLegend}
\usage{
plotSpeciesLegend(pos = "top", ...)
}
\description{
This function add the species legend to a plot
 @param pos the legend position on plot area
}

