% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{plotToFile}
\alias{plotToFile}
\title{plotToFile}
\usage{
plotToFile(file, plotFunc, ...)
}
\arguments{
\item{file}{the file to export the plot}

\item{plotFunc}{to be exported}
}
\description{
This function export to a file the original PLot
}

