% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.com.R
\name{qboxplot}
\alias{qboxplot}
\title{qboxplots for matrix columns or list of vectors}
\usage{
qboxplot(vals, labs = NULL, lims = NULL, whiskerQuantile = 0.05,
  horizontal = F, ...)
}
\arguments{
\item{vals}{a dataset as list or matrix of values}

\item{labs}{a vector of names for each dataset}

\item{lims}{overrides yLim parameter of boxplot}

\item{whiskerQuantile}{whiskers range from (this value) to (1 - this value)}
}
\value{
a boxplot
}
\description{
qboxplots for matrix columns or list of vectorsusing controllable quantile based whiskers
}

