% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{saveLogRatios}
\alias{saveLogRatios}
\title{saveLogRatios}
\usage{
saveLogRatios(rs)
}
\arguments{
\item{rs}{the result set}
}
\description{
extract and store calculated log-ratios from a list of result sets
}

