% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{saveMetrics}
\alias{saveMetrics}
\title{saveMetrics}
\usage{
saveMetrics(resultSets = ResultSets, File = cfg$LogFilesLocation +
  "/metrics.txt")
}
\arguments{
\item{resultSets}{}

\item{File}{}
}
\description{
This function save all the metrics related with the experiment
}

