% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.access.R
\name{saveMetrics}
\alias{saveMetrics}
\title{saveMetrics}
\usage{
saveMetrics(resultSets = ResultSets, File = cfg$LogFilesLocation +
  "/metrics.txt")
}
\arguments{
\item{File}{the file to export the current metrics}

\item{resulSets}{the result List from the analysis}
}
\description{
This function save the computed metrics to an output file
}

