% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.access.r
\name{saveSpeciesSeparation}
\alias{saveSpeciesSeparation}
\title{saveSpeciesSeparation}
\usage{
saveSpeciesSeparation(resultSet)
}
\arguments{
\item{resultSet}{}
}
\description{
This function enable to save the species
}

