% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{showDistributionDensityPlot}
\alias{showDistributionDensityPlot}
\title{showdistributionDensityPlot}
\usage{
showDistributionDensityPlot(samplePair, showLegend = F)
}
\arguments{
\item{samplePair}{}

\item{showLegend}{}
}
\description{
This fucntion draw log-ratio distribution kernel density plot
}

