% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{showLogRatioBoxPlot}
\alias{showLogRatioBoxPlot}
\title{showLogRatioBoxPlot}
\usage{
showLogRatioBoxPlot(samplePair)
}
\arguments{
\item{samplePair}{the result set data of a hybrid proteome sample pair}
}
\description{
draw log ratio quiantile based boxplot
}

