% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{showQuantBarPlot}
\alias{showQuantBarPlot}
\title{showquantBarPlot
This function draw barplot showing quantification bars}
\usage{
showQuantBarPlot(samplePair)
}
\arguments{
\item{samplePair}{}
}
\description{
showquantBarPlot
This function draw barplot showing quantification bars
}

