% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfqbench.plot.r
\name{showSingleSpeciesQC}
\alias{showSingleSpeciesQC}
\title{showSingleSpeciesQC}
\usage{
showSingleSpeciesQC(samplePair, speciesName = "MOUSE")
}
\arguments{
\item{samplePair}{the result set data of a hybrid proteome sample pair}

\item{speciesName}{the name of a species}
}
\description{
draw quantification curves
}

