% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.plot.R
\name{showSingleSpeciesQC}
\alias{showSingleSpeciesQC}
\title{showSingleSpeciessQC}
\usage{
showSingleSpeciesQC(samplePairsData, speciesName = "MOUSE")
}
\arguments{
\item{samplePairsData}{}

\item{speciesName}{}
}
\description{
This function draw the QC values for one particular specie
}

