% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqbench.com.r
\name{spiner}
\alias{spiner}
\title{spiner}
\usage{
spiner(samples, sampleNames = colnames(samples),
  partNames = rownames(samples), plotTitle = "", bgCols = NULL,
  fgCols = "black")
}
\arguments{
\item{sampleNames}{}

\item{partNames}{}

\item{plotTitle}{}

\item{fgCols}{}
}
\description{
This function compute the spiner
}

