% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFQbench.com.R
\name{spiner}
\alias{spiner}
\title{spiner}
\usage{
spiner(samples, sampleNames = colnames(samples),
  partNames = rownames(samples), plotTitle = "", bgCols = NULL,
  fgCols = "black")
}
\arguments{
\item{samples}{}

\item{sampleNames}{samples names}

\item{partNames}{Row names}

\item{plotTitle}{plot Title}
}
\value{
spineplot
}
\description{
generate an spineplot using different paramters. Spineplot are a special cases of mosaic plots, and can be seen as a generalization of stacked (or highlighted) bar plots. Analogously, spinograms are an extension of histograms.
}

